+++ src/canvas/colorcanvas.h
@@ -13,7 +13,7 @@ namespace PixelMaestro {
 			ColorCanvas(Section* section);
 			ColorCanvas(Section *section, uint16_t num_frames);
 			~ColorCanvas();
-			Colors::RGB get_pixel_color(uint32_t pixel);
+			Colors::RGB get_pixel_color(uint16_t x, uint16_t y);
 
 			// Overriden functions
 			void activate(uint32_t pixel);